tutoria 08: ejercicio 8

+++ Utilities/Validations.cs
@@ -2,13 +2,13 @@
 
 public class Validations
 {
-    public static double ValidateNumber(string input)
+    public static int ValidateNumber(string input)
     {
         //declaracion + inicializacion de variable
-        double result = 0;
+        int result = 0;
 
         // validacion 
-        if (!double.TryParse(input, out result))
+        if (!int.TryParse(input, out result))
         {
             Console.WriteLine("Por favor ingrese un valor numerico");
             return 0;
